backend/insta_backend/requirements.txt
amqp==5.0.6
anyjson==0.3.3
apsw==3.28.0.post1
apturl==0.5.2
asgiref==3.3.1
asn1crypto==0.24.0
attrs==19.3.0
Automat==0.8.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==3.6.4.0
blinker==1.4
breezy==3.0.2
Brlapi==0.7.0
celery==5.0.5
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
css-parser==1.0.4
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
deluge==2.0.3
Deprecated==1.2.7
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
dnspython==1.16.0
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
empy==3.3.2
entrypoints==0.3
fasteners==0.14.1
fastimport==0.9.8
feedparser==5.2.1
future==0.18.2
GeoIP==1.3.2
gpg===1.13.1-unknown
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
ifaddr==0.1.6
img2pdf==0.3.3
incremental==16.10.1
keyring==18.0.1
keyrings.alt==3.4.0
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
mechanize==0.4.5
monotonic==1.5
msgpack==0.6.2
netifaces==0.10.4
nose==1.3.7
numpy==1.17.4
oauth==1.0.1
oauthlib==3.1.0
ocrmypdf==9.6.0+dfsg
olefile==0.46
openshot-qt==2.4.3
packaging==20.3
PAM==0.4.2
paramiko==2.6.0
pbr==5.4.5
pdfminer.six==20191020
pexpect==4.6.0
pikepdf==1.10.3+dfsg
Pillow==7.0.0
prompt-toolkit==3.0.18
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.8.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pychm==0.8.6
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.6
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyPDF2==1.26.0
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.17.1
python-gitlab==2.0.1
python-libtorrent==1.1.13
python-magic==0.4.16
python-pam==1.8.4
python-xapp==1.8.1
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
redis==3.2.0
regex==2019.8.19
rencode==1.0.6
reportlab==3.5.34
repoze.lru==0.7
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
Routes==2.4.1
ruffus==2.8.1
scour==0.37
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sortedcontainers==2.1.0
soupsieve==1.9.5
sqlparse==0.4.1
ssh-import-id==5.10
system-service==0.3
systemd-python==234
testresources==2.0.0
tinycss==0.4
tqdm==4.30.0
traceback2==1.4.0
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.25.8
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.5
whitenoise==5.2.0
wrapt==1.11.2
xkit==0.0.0
zeroconf==0.24.4
zope.interface==4.7.1
